logic duplicate clean-up 1/2: makeRoomForClipsOnTracks and makeRoomForDataOnTracks co-exist in Au3TracksInteraction
