+++ src/trackedit/internal/au3/au3clipsinteraction.cpp
@@ -744,7 +744,7 @@ muse::Ret Au3ClipsInteraction::makeRoomForClip(const ClipKey& clipKey)
             continue;
         }
 
-        trimOrDeleteOverlapping(waveTrack, clip->GetPlayStartTime(), clip->GetPlayEndTime(), otherClip);
+        utils::trimOrDeleteOverlapping(prj, waveTrack, clip->GetPlayStartTime(), clip->GetPlayEndTime(), otherClip);
     }
 
     IF_ASSERT_FAILED(noPlayRegionsOverlap(clipKey.trackId)) {
@@ -1382,122 +1382,4 @@ bool Au3ClipsInteraction::singleClipOnTrack(const TrackId trackId) const
     }
 
     return waveTrack->Intervals().size() == 1;
-}
-
-void Au3ClipsInteraction::trimOrDeleteOverlapping(::WaveTrack* waveTrack,
-                                                  muse::secs_t begin,
-                                                  muse::secs_t end,
-                                                  std::shared_ptr<::WaveClip> otherClip)
-{
-    trackedit::ITrackeditProjectPtr prj = globalContext()->currentTrackeditProject();
-
-    if (muse::RealIsEqualOrLess(begin, otherClip->GetPlayStartTime())
-        && muse::RealIsEqualOrMore(end, otherClip->GetPlayEndTime())) {
-        auto clipInfo = DomConverter::clip(waveTrack, otherClip.get());
-        waveTrack->RemoveInterval(otherClip);
-        prj->notifyAboutClipRemoved(clipInfo);
-        prj->notifyAboutTrackChanged(DomConverter::track(waveTrack));
-        return;
-    }
-
-    if (!muse::RealIsEqualOrLess(begin, otherClip->GetPlayStartTime())
-        && !muse::RealIsEqualOrMore(end, otherClip->GetPlayEndTime())) {
-        muse::secs_t otherClipStartTime = otherClip->GetPlayStartTime();
-        muse::secs_t otherClipEndTime = otherClip->GetPlayEndTime();
-
-        auto leftClip = waveTrack->CopyClip(*otherClip, true);
-        waveTrack->InsertInterval(std::move(leftClip), false);
-        prj->notifyAboutTrackChanged(DomConverter::track(waveTrack));
-
-        muse::secs_t rightClipOverlap = (end - otherClip->GetPlayStartTime());
-        otherClip->TrimLeft(rightClipOverlap);
-        prj->notifyAboutClipChanged(DomConverter::clip(waveTrack, otherClip.get()));
-
-        leftClip->SetPlayStartTime(otherClipStartTime);
-        muse::secs_t leftClipOverlap = (otherClipEndTime - begin);
-        leftClip->TrimRight(leftClipOverlap);
-        prj->notifyAboutClipAdded(DomConverter::clip(waveTrack, leftClip.get()));
-
-        prj->notifyAboutTrackChanged(DomConverter::track(waveTrack));
-        return;
-    }
-
-    if (muse::RealIsEqualOrLess(begin, otherClip->GetPlayStartTime())
-        && !muse::RealIsEqualOrMore(end, otherClip->GetPlayEndTime())
-        && muse::RealIsEqualOrMore(end, otherClip->GetPlayStartTime())) {
-        muse::secs_t overlap = (end - otherClip->GetPlayStartTime());
-        otherClip->TrimLeft(overlap);
-        prj->notifyAboutClipChanged(DomConverter::clip(waveTrack, otherClip.get()));
-        return;
-    }
-
-    if (!muse::RealIsEqualOrLess(begin, otherClip->GetPlayStartTime())
-        && muse::RealIsEqualOrLess(begin, otherClip->GetPlayEndTime())
-        && muse::RealIsEqualOrMore(end, otherClip->GetPlayEndTime())) {
-        muse::secs_t overlap = (otherClip->GetPlayEndTime() - begin);
-        otherClip->TrimRight(overlap);
-        prj->notifyAboutClipChanged(DomConverter::clip(waveTrack, otherClip.get()));
-        return;
-    }
-}
-
-muse::Ret Au3ClipsInteraction::makeRoomForDataOnTrack(const TrackId trackId,
-                                                      muse::secs_t begin,
-                                                      muse::secs_t end)
-{
-    trackedit::ITrackeditProjectPtr prj = globalContext()->currentTrackeditProject();
-
-    WaveTrack* waveTrack = DomAccessor::findWaveTrack(projectRef(), ::TrackId(trackId));
-    IF_ASSERT_FAILED(waveTrack) {
-        return make_ret(trackedit::Err::TrackNotFound);
-    }
-
-    std::list<std::shared_ptr<WaveClip> > clips = DomAccessor::waveClipsAsList(waveTrack);
-
-    clips.sort([](const std::shared_ptr<WaveClip>& c1, const std::shared_ptr<WaveClip>& c2) {
-        return c1->GetPlayStartTime() < c2->GetPlayStartTime();
-    });
-
-    for (const auto& otherClip : clips) {
-        trimOrDeleteOverlapping(waveTrack, begin, end, otherClip);
-    }
-
-    return make_ret(muse::Ret::Code::Ok);
-}
-
-muse::Ret Au3ClipsInteraction::makeRoomForClipsOnTracks(const std::vector<TrackId>& tracksIds,
-                                                        const std::vector<ITrackDataPtr>& trackData,
-                                                        muse::secs_t begin)
-{
-    trackedit::ITrackeditProjectPtr prj = globalContext()->currentTrackeditProject();
-
-    IF_ASSERT_FAILED(tracksIds.size() <= trackData.size()) {
-        return make_ret(trackedit::Err::NotEnoughDataInClipboard);
-    }
-
-    for (size_t i = 0; i < tracksIds.size(); ++i) {
-        WaveTrack* dstWaveTrack = DomAccessor::findWaveTrack(projectRef(), ::TrackId(tracksIds.at(i)));
-        IF_ASSERT_FAILED(dstWaveTrack) {
-            return make_ret(trackedit::Err::TrackNotFound);
-        }
-
-        muse::secs_t snappedBegin = dstWaveTrack->SnapToSample(begin);
-
-        auto au3TrackData = std::dynamic_pointer_cast<Au3TrackData>(trackData.at(i));
-        if (!au3TrackData) {
-            continue;
-        }
-
-        const WaveTrack* wt = dynamic_cast<const Au3WaveTrack*>(au3TrackData->track().get());
-        for (const auto& interval : wt->Intervals()) {
-            auto ok = makeRoomForDataOnTrack(tracksIds.at(i),
-                                             snappedBegin + interval->GetPlayStartTime(),
-                                             snappedBegin + interval->GetPlayEndTime());
-            if (!ok) {
-                return make_ret(trackedit::Err::FailedToMakeRoomForClip);
-            }
-        }
-    }
-
-    return muse::make_ok();
 }
