+++ src/trackedit/internal/au3/au3interactionutils.cpp
@@ -7,11 +7,13 @@
 
 #include "../../trackedittypes.h"
 #include "au3wrap/internal/domaccessor.h"
+#include "au3wrap/internal/domconverter.h"
 #include "au3-wave-track/WaveTrack.h"
 #include "au3-wave-track/WaveClip.h"
 #include "au3-project-rate/ProjectRate.h"
 #include "au3-project-rate/QualitySettings.h"
 #include "global/containers.h"
+#include "global/realfn.h"
 #include "log.h"
 #include <unordered_set>
 
@@ -94,7 +96,7 @@ size_t au::trackedit::utils::getTrackIndex(const au3::Au3TrackList& tracks, cons
 
 void au::trackedit::utils::exchangeTrack(au3::Au3TrackList& tracks, au3::Au3WaveTrack& oldOne, au3::Au3WaveTrack& newOne)
 {
-    auto tmp = TrackList::Temporary(nullptr, newOne.shared_from_this());
+    auto tmp = au3::Au3TrackList::Temporary(nullptr, newOne.shared_from_this());
     tracks.ReplaceOne(oldOne, std::move(*tmp));
 }
 
@@ -310,4 +312,57 @@ muse::Ret au::trackedit::utils::withProgress(muse::IInteractive& interactive, co
     progress.finish(muse::make_ok());
 
     return result;
+}
+
+void au::trackedit::utils::trimOrDeleteOverlapping(const ITrackeditProjectPtr& project, au3::Au3WaveTrack* waveTrack,
+                                                   secs_t begin, secs_t end, std::shared_ptr<au3::Au3WaveClip> otherClip)
+{
+    if (muse::RealIsEqualOrLess(begin, otherClip->GetPlayStartTime())
+        && muse::RealIsEqualOrMore(end, otherClip->GetPlayEndTime())) {
+        auto clipInfo = au::au3::DomConverter::clip(waveTrack, otherClip.get());
+        waveTrack->RemoveInterval(otherClip);
+        project->notifyAboutClipRemoved(clipInfo);
+        project->notifyAboutTrackChanged(au::au3::DomConverter::track(waveTrack));
+        return;
+    }
+
+    if (!muse::RealIsEqualOrLess(begin, otherClip->GetPlayStartTime())
+        && !muse::RealIsEqualOrMore(end, otherClip->GetPlayEndTime())) {
+        secs_t otherClipStartTime = otherClip->GetPlayStartTime();
+        secs_t otherClipEndTime = otherClip->GetPlayEndTime();
+
+        auto leftClip = waveTrack->CopyClip(*otherClip, true);
+        waveTrack->InsertInterval(std::move(leftClip), false);
+        project->notifyAboutTrackChanged(au::au3::DomConverter::track(waveTrack));
+
+        secs_t rightClipOverlap = (end - otherClip->GetPlayStartTime());
+        otherClip->TrimLeft(rightClipOverlap);
+        project->notifyAboutClipChanged(au::au3::DomConverter::clip(waveTrack, otherClip.get()));
+
+        leftClip->SetPlayStartTime(otherClipStartTime);
+        secs_t leftClipOverlap = (otherClipEndTime - begin);
+        leftClip->TrimRight(leftClipOverlap);
+        project->notifyAboutClipAdded(au::au3::DomConverter::clip(waveTrack, leftClip.get()));
+
+        project->notifyAboutTrackChanged(au::au3::DomConverter::track(waveTrack));
+        return;
+    }
+
+    if (muse::RealIsEqualOrLess(begin, otherClip->GetPlayStartTime())
+        && !muse::RealIsEqualOrMore(end, otherClip->GetPlayEndTime())
+        && muse::RealIsEqualOrMore(end, otherClip->GetPlayStartTime())) {
+        secs_t overlap = (end - otherClip->GetPlayStartTime());
+        otherClip->TrimLeft(overlap);
+        project->notifyAboutClipChanged(au::au3::DomConverter::clip(waveTrack, otherClip.get()));
+        return;
+    }
+
+    if (!muse::RealIsEqualOrLess(begin, otherClip->GetPlayStartTime())
+        && muse::RealIsEqualOrLess(begin, otherClip->GetPlayEndTime())
+        && muse::RealIsEqualOrMore(end, otherClip->GetPlayEndTime())) {
+        secs_t overlap = (otherClip->GetPlayEndTime() - begin);
+        otherClip->TrimRight(overlap);
+        project->notifyAboutClipChanged(au::au3::DomConverter::clip(waveTrack, otherClip.get()));
+        return;
+    }
 }
